Module twitch_api_rs::auth

Types and functions for auth flows

See Twitch Auth Documentation

Modules

authorization_code
Authorization Code Flow

client_credentials
Client Credentials Flow

implicit_code
Implicit Code Flow

scopes
Types relating to scopes that a Bearer Auth token can have
Used with crate::auth::AuthToken

Structs

ClientId
Represents a crate::auth::client_credentials id.
See Twitch Auth Guide for more

ClientSecret
Represents a crate::auth::client_credentials secret.
See Twitch Auth Guide for more

Traits

AuthToken
Represents a authorization token of some type that can be sent as a header to a twitch endpoint.