Trait twitch_api_rs::auth::AuthToken

pub trait AuthToken: Headers + Clone {
    fn scopes(&self) -> &ScopeSet;
}

Represents a authorization token of some type that can be sent as a header to a twitch endpoint.

Required methods

fn scopes(&self) -> &ScopeSet

Get the set of scopes that this token has associated with it

Loading content...