Module twitch_api_rs::requests

Common traits used to construct requestable types

TODO: Make make_request function based on feature and requests constructable without async or sending in all cases

Structs

FailureStatus
Represents a sucessful request that was denied by the twitch api for some reason. Use request's associated ErrorCodes to get matchable value.

Enums

CommonResponseCodes
Error codes used by twitch that are the same across most endpoints.

PossibleResponse
Represents a possible response from the twitch api, deserialized from a sucessful request. May not contain the requested content but instead a FailureStatus

RequestError
Returned from a request when it could not be completed

Traits

Body
Body for a request

BodyExt
Marker trait for auto implementation of Body for types that implement serde::Serialize

ErrorCodes
Represents A Known set of error status codes that an endpoint may return.o

Headers
Headers for a request

HeadersExt
Marker trait for auto implementation of headers

Parameters
Parameters for a request

ParametersExt
Marker trait for auto implementation of Parameters for types that implement serde::Serialize

Request
Represents a request that can be made to the twitch api

Response
Type that is returned by a sucessful request