Module twitch_api_rs::values

Wrappers for basic owned types that indicate their usage

Modules

broadcasters
Values for broadcaster objects and requests

clips
Values for clip objects and requests

extensions
Values for extension objects and requests

games
Values for game objects and requests

users
Values for user objects and requests

videos

Structs

Count
The max amount returned per page, used in requests like crate::resource::clips::get_clips::GetClipsRequest

EndedAt
Represents the end of a time window

ISOLanguage
Represents a language, either a ISO 639-1 two-letter language code or 'other'

Pagination
A Pagination key for enpoints that may return more than 100 results

Period
Represents a time window

RFC3339Time
Represents a RFC3339 formatted datetime

StartedAt
Represents the beginning of a time window

Url

Traits

FieldValue
Used to indicated that this type is used a field value