Trait twitch_api_rs::values::FieldValue

pub trait FieldValue {
    fn field_name() -> &'static str;
}

Used to indicated that this type is used a field value

Required methods

fn field_name() -> &'static str

Get the commonly used name of a field of this type that twitch is expecting

Loading content...